Simplify error handling and event passing
We introduce 2 new exceptions:

- InvalidMessageException, to be used when a message doesn't conform to the IRC
  message syntax, and

- UnexpectedMessageException, to be used when a message of a kind different than
  expected was received.

IrcChannel's com() method no longer returns null in case of failures such as
above, but throws an appropriate exception.

Furthermore, we no longer constrain NickMessage, UserMessage, etc. to be
"correct by construction", in the sense that their constructors and construct()
methods would not allow invalid instances to be constructed -- either by making
it impossible through the constructor argument list, or by throwing exceptions.
Previously, IrcChannel's com() would propagate such exceptions and they would
have to be caught in the choreography, which is pretty ugly.

Now we allow NickMessage, UserMessage, etc. to be "invalid", and their usage
boils down to being just a convenience wrapper around a generic Message.
However, now we explicitly check the validity of the messages in the
choreography and report an appropriate error message.