Remove UnexpectedMessageException
Because of the way Java implements type erasure, the cast to the generic type
parameter is not done within the implementation of com(), but within the
particular call site of com(). In other words, our ClassCastException handler is
completely useless, and unless we want to wrap up every call site of com(), so
is UnexpectedMessageException.