wip: Try and catch ClassCastException
In the end, catching a ClassCastException that's the result of a mismatch
between what com() expects and what is deserialized just isn't possible.

Because com() is generic and Java uses type erasure, a cast to a type parameter
is not actually done within the body of com(), but at every call site that
invokes com(). The cast is automatically generated by the Java compiler.

Even if we push the generic type parameter into Message.construct and try to use
various Class tricks (Class.cast(), Class.asSubclass(), etc.), we're still left
with the same problem in the end as there's no way for us to ever get an
instance of Class<M>. The best we can do is Class<? extends Message> which is
not enough.