Bump Java version to 19
In order to make use of Java's --enable-preview flag, the Java version requested
on the command-line (or indirectly through the POM) has to match the one of the
used compiler.

This will end up in a broken build as soon as the requested version is lower
than the version of the user's compiler, so let's just keep the requested
version the highest it can be (and push it onto users to keep the compiler up to
date).