Don't use addServerMessage internally for consistency
We are not able to use addClientMessage() in the one case where the client adds
something to its queue, so let's not use it for the server either.

The only reason we're able to use it at the server in this case is because all
branches execute the same action, addServerMessage(). If any branch did
something different, this would result in a merging error.

The reason this is so is that a method of a multi-role choreography is treated
as if it necessarily involves all of the choreography's participants, whether or
not it actually does so. Using such a method within an if that's supposed to be
completely local to one role turns it into a "choreographic if", and therefore
pulls in the problem of knowledge of choice, selection, merging, etc.