Avoid the race condition in stress-client
It turns out that killing the sleep process which holds the pipe open can lead
to Netcat receiving an EOF and terminating before it has received all of the
data from the server. In that case, an exception is raised at the server since
the connection was closed ungracefully, while the server was trying to send data.

We now only kill the sleep process after we receive an ERROR IRC message from
the server. This is good enough since the server will surely have closed the
connection after that (or perhaps there's yet another race condition here?).

It would be so much easier if Netcat would just terminate once the server closes
its connection, but it seems this behavior varies between Netcat
implementations, and even different versions of the same implementation.

Modern versions of Netcat prefer to remain running in a half-duplex state, and
I'm not sure what's the best way to make them terminate once the remote side
closes its connection. Using Ncat's --no-shutdown option doesn't help.