Use Bash's /dev/tcp instead of Netcat
Even with killing the sleep process and closing the pipe only once we get an
ERROR IRC message, a new issue appeared where the connection is prematurely
closed by the client (the server is waiting to receive from the client and fails
with a connection closed error).

The issue occurs every once in a while when stress testing, with as few as 3
clients. Further, it occurs even though the sleep process is still alive and the
pipe open. Even worse, Netcat is for some reason not reading the pipe fully,
which can be confirmed by manually reading from the pipe in a different shell.
I'm not sure why any of this is happening, especially why Netcat is not reading
input from the pipe.

To get around this whole issue, we replace Netcat with Bash's /dev/tcp which
seems to work fine and terminate properly so far.