Immediately unregister a client on QUIT
The client is unregistered immediately after receiving a QUIT message, rather
than once the connection is closed. This is necessary so that any actions and
messages received afterward observe that the nickname has been released and that
the client has left all channels. We do not want to have a race condition and
depend on how long it takes to close the connection.

The client entry is still kept around in the server's state and is actually
removed once the connection closes.